% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotSquare.R
\name{plotSquare}
\alias{plotSquare}
\title{Plot a square, centered on the origin, of radius r.}
\usage{
plotSquare(r = 1)
}
\arguments{
\item{r}{The "half-width" of the square to be plotted}
}
\value{
None The function is called solely for its
  side effect
}
\description{
This plots a square.
The "radius" is the half-width.
}
\examples{
plotSquare(10)
}
\seealso{
The general \code{\link[graphics]{plot}} function
}

