% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vectorOfRadii.R
\name{vectorOfRadii}
\alias{vectorOfRadii}
\title{A bunch of radii (TITLE)}
\format{a description of what the variables are,
  with full names and explanations (especially
  if this is a data frame). Dimensions help.}
\source{
where did the data come from? This could
  be a URL, or some other type of pointer.
}
\description{
Really, do you need this explained?
Well, ok. (DESCRIPTION)
}

